pedalkernel-spice-harness/requirements.txt
numpy==1.26.4
scipy==1.13.1
matplotlib==3.9.2
soundfile==0.12.1
pyyaml==6.0.2
jsonschema==4.23.0
click==8.1.7
tabulate==0.9.0
PySpice==1.5
jinja2==3.1.4
